Moved rack-logger and changed require statements
