+++ lib/a2b_logging/rails_ext/rack/logger.rb
